+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file. (http://keepachangelog.com/)
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## unreleased
+- Installed Sentry
+
 ## [1.0.6] - 2010-06-02
 - Fixed error page